Minor optimization
